feat(service): Assign supportgroup to unknown service (#629)
The k8s-assets scanner is reading owner
information from the owner-helm-chart.
The supportGroup attribute is mandatory, but
the service attribute is optional.
Currently, if the scanner encounters such a case,
it adds the componentInstance to a None service.
This will lead to one none service, that has
componentInstances assigned from multiple
supportGroup including all their vulnerabilities.

Instead of a single 'none` service for all Support
Groups, we want to have one per Support Group.
The name for this will be "Unknown-{SupportGroupName}"