db(refactor): remove TestConnection
TestConnection function seems to be not needed
it is checking db connection in every test, but instead
of calling TestConnection we are setting dbSchema using
exec USE <name> it will fail if connection could not be
established. Furthermore TestConnection has some drawbacks.
First it is testing sql DB connection if no DBName is not
provided, but if DBName is provided it is used to check
if DBName exists, this clearly brakes SRP rule. The other
issue with TestConnection function is that it is probably
used to wait for db to be ready, trying to reach it every
defined interval. So there is third kind of usage for the
function. Furthermore it is recursive function, recursive
functions obscures call stack and makes debugging troubles.
Considering mentioned drawbacks and the minor advantage of
using this function in testing it is not worth to keep it
in the code.

On-behalf-of: SAP Michal Krzyz <michal.krzyz@sap.com>
Signed-off-by: Michal Krzyz <michalkrzyz@gmail.com>